requirements.txt
# Binance API
binance-futures-connector>=4.0.0
python-binance>=1.0.19

# Data & Technical Analysis
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
yfinance>=0.2.28

# News & Sentiment (optional: cần API key)
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# AI/ML - Phân tích & dự đoán
scikit-learn>=1.3.0
xgboost>=2.0.0

# Config & Utils
python-dotenv>=1.0.0
schedule>=1.2.0
loguru>=0.7.0

# Dashboard web
flask>=3.0.0

# Async (optional)
aiohttp>=3.9.0
